Daemon: Add protocol to provide buildsystem information for a file
This should provide the compile definitions, include directories etc
as JSON properties, but should also provide the full command line
content (including ordering of include directories, content of
CMAKE_CXX_FLAGS etc) which gets passed to the compiler, if that is
possible to do.

TODO: Refactor CMake to provide this information.