cmake: Issue message independent of cmMakefile definition
The makefile is only used when called by the cmMessageCommand, so inline
the use of it there.  It creates an undesirable dependency on
cmMakefile for issuing messages in the cmake instance.

This also makes it more explicit that the variable definitions only
affect the message() command.  If an AUTHOR_WARNING is issued for any
other reason, it is not affected.  To affect that, it is necessary to
set the cache variable instead of the regular variable.

This is an unfortunate interface quirk.  Perhaps for newly introduced
variables, the non-cache version should be ignored to encourage
consistency of setting the cache variables instead.